+++ be/src/column/json_column.h
@@ -42,7 +42,7 @@ public:
     int compare_at(size_t left, size_t right, const starrocks::Column& rhs, int nan_direction_hint) const override;
     void fnv_hash(uint32_t* hash, uint32_t from, uint32_t to) const override;
     void put_mysql_row_buffer(starrocks::MysqlRowBuffer* buf, size_t idx) const override;
-    std::string debug_item(uint32_t idx) const override;
+    std::string debug_item(size_t idx) const override;
     std::string get_name() const override;
 
     const uint8_t* deserialize_and_append(const uint8_t* pos) override;
