+++ be/src/column/object_column.cpp
@@ -289,17 +289,17 @@ ColumnPtr ObjectColumn<T>::clone_shared() const {
 }
 
 template <typename T>
-std::string ObjectColumn<T>::debug_item(uint32_t idx) const {
+std::string ObjectColumn<T>::debug_item(size_t idx) const {
     return "";
 }
 
 template <>
-std::string ObjectColumn<HyperLogLog>::debug_item(uint32_t idx) const {
+std::string ObjectColumn<HyperLogLog>::debug_item(size_t idx) const {
     return _pool[idx].to_string();
 }
 
 template <>
-std::string ObjectColumn<BitmapValue>::debug_item(uint32_t idx) const {
+std::string ObjectColumn<BitmapValue>::debug_item(size_t idx) const {
     return _pool[idx].to_string();
 }
